Modelled other classes (mage, cleric, ranger, necromancer) after warrior and tested in main to see if they work properly.

+++ src/stat_type.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-typedef std::int8_t stat_t; // 0 - 255
+typedef std::int16_t stat_t; // 0 - 255
